Section B: Removed some commented lines; Added boolean_replacement function which replaces boolean columns by "Sim" and "Não" and vice-versa;
Section E:
Added an argument to sql_truncate function: query. If query is none, it truncates the whole view, otherwise it executes the provided query;

Streamlit Reports - Optionals Baviera Solver:
This reports not only uses an optimizer on the fly to solve for the best group of vehicles, but also checks for the currently saved pairs of Sales Place/Model. If a pair already exists, it asks the user if he wants to overwrite the solution. In order to implement this, a small gist (sessionstate) was used where Session State is saved between different runs of the report; The suggestion number for each configuration is given by the ratio of the score of said configuration in the overall group of configurations selected by the solver. Also, the variability is forced manually by selecting the configuration with the value of the parameter selected by the user with the highest score and forcing such configuration to have a value in the final solution, greater than 1;

Order Suggestion Baviera Options:
Replaced the call to the module neighbors from sklearn to account for FutureWarnings; Added the optimization_solution_table to sql_info;

Order Suggestion Hyundai Options:
Replaced the call to the module neighbors from sklearn to account for FutureWarnings; Corrected the call to the module svm from sklearn;

+++ modules/level_1_e_deployment.py
@@ -245,10 +245,14 @@ def sql_string_preparation_v1(values_list):
     return columns_string, values_string
 
 
-def sql_truncate(dsn, options_file, database, view):
+def sql_truncate(dsn, options_file, database, view, query=None):
     level_0_performance_report.log_record('A truncar view {} da DB {}.'.format(view, database), options_file.project_id)
+
     cnxn = pyodbc.connect('DSN={};UID={};PWD={};DATABASE={}'.format(dsn, options_file.UID, options_file.PWD, database), searchescape='\\')
-    query = "TRUNCATE TABLE " + view
+
+    if query is None:
+        query = "TRUNCATE TABLE " + view
+
     cursor = cnxn.cursor()
     cursor.execute(query)
